+++ satella/instrumentation/memory/__init__.py
@@ -2,7 +2,9 @@ from .conditions import Any, All, GlobalRelativeValue, GlobalAbsoluteValue, Loca
     LocalRelativeValue, GB, MB, KB, CustomCondition, Not
 from .default import install_force_gc_collect
 from .memthread import MemoryPressureManager
+from .dump_frames_on import dump_memory_on, install_dump_memory_on
 
 __all__ = ['Any', 'All', 'MemoryPressureManager', 'GlobalAbsoluteValue',
            'GB', 'GlobalRelativeValue', 'LocalRelativeValue', 'LocalAbsoluteValue', 'MB', 'KB',
-           'CustomCondition', 'Not', 'install_force_gc_collect']
+           'CustomCondition', 'Not', 'install_force_gc_collect',
+           'dump_memory_on', 'install_dump_memory_on']