+++ CHANGELOG.md
@@ -1,2 +1,4 @@
 # v2.17.13
 
+* added memory dumping
+* exposed `dump_frames_on`